ProductInventory/requirements.txt
Django==4.2.6
django-admin==2.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.2.0
gunicorn==21.2.0
